+++ extension/package.json
@@ -2,7 +2,7 @@
   "name": "ducklab",
   "displayName": "Ducklab (preview)",
   "description": "Interactive SQL Notebooks. Powered by DuckDB",
-  "version": "0.3.2",
+  "version": "0.4.0",
   "license": "Apache-2.0",
   "publisher": "ducklake",
   "icon": "./logo.png",
